Partial progress on upgrading to v5.5 driver
driver/i2s.h is deprecated and significantly more limited especially wrt
handling overrun/underrun. However, I'm having a lot of trouble getting
it to be stable and clean, must be some bad config settings somewhere,
but I'll come back to it. This code state is stripped down for testing
but I don't expect there to be module-specific changes required, so this
can be upgraded whenever.